% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{annoEtranger.rda}
\alias{annoEtranger}
\alias{annoEtranger.rda}
\title{Annotation of first two lines of Albert Camus' L'Etranger}
\format{a \code{annotation} object}
\usage{
annoEtranger
}
\description{
Parsed via the Stanford CoreNLP Java Library
}
\author{
Taylor Arnold, 2015-06-03
}
\keyword{datasets}

