% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{annoHp.rda}
\alias{annoHp}
\alias{annoHp.rda}
\title{Annotation of first line of JK Rowling's The Philosopher's Stone}
\format{a \code{annotation} object}
\usage{
annoHp
}
\description{
Parsed via the Stanford CoreNLP Java Library
}
\author{
Taylor Arnold, 2015-06-03
}
\keyword{datasets}

