% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp.R
\name{annotateFile}
\alias{annotateFile}
\title{Annotate a text file}
\usage{
annotateFile(file, format = c("obj", "xml", "text"), outputFile = NA,
  includeXSL = FALSE)
}
\arguments{
\item{file}{a string giving the location of the file to be loaded.}

\item{format}{the desired output format. Option \code{obj}, the default,
returns an R object of class \code{annotation} and will
likely be the desired choice for users loading the output
into R. The \code{xml} and \code{text} exist primarily for
saving the files on the disk.}

\item{outputFile}{character string indicating where to put the output. If
set to NA, the output will be returned by the function.}

\item{includeXSL}{boolean. Whether the xml style sheet should be included
in the output. Only used if format is \code{xml} and
outputFile is not \code{NA}.}
}
\description{
Runs the CoreNLP annotators for the text contained in a given file.
The details for which annotators to run and how to run them are
specified in the properties file loaded in via the \code{initCoreNLP}
function (which must be run prior to any annotation).
}

