% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/download.R
\name{downloadCoreNLP}
\alias{downloadCoreNLP}
\title{Download java files needed for coreNLP}
\usage{
downloadCoreNLP(outputLoc, type = c("base", "chinese", "spanish", "sr"))
}
\arguments{
\item{outputLoc}{a string showing where the files are to be downloaded.
If missing, will try to download files into the directory
where the package was original installed.}

\item{type}{type of files to download. The base backage, installed by
default is required. Other jars include chinese, spanish,
and "shift reduce parser models". These will be installed
in addition to the base package.}
}
\description{
The coreNLP package does not supply the raw java files
provided by the Stanford NLP Group as they are quite large.
This function downloads the libraries for you, by default into
the directory where the package was installed.
}
\examples{
\dontrun{
downloadCoreNLP()
downloadCoreNLP(type="spanish")
}
}

