% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp.R
\name{getCoreference}
\alias{getCoreference}
\title{Get Coreference}
\usage{
getCoreference(annotation)
}
\arguments{
\item{annotation}{an annotation object}
}
\description{
Returns a dataframe containing all coreferences detected in the text.
}
\examples{
getCoreference(annoHp)

}

