% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/nlp.R
\name{getDependency}
\alias{getDependency}
\title{Get Coreferences}
\usage{
getDependency(annotation, type = c("basic", "collapsed", "CCprocessed"))
}
\arguments{
\item{annotation}{an annotation object}

\item{type}{the class of coreference desired}
}
\description{
Returns a data frame of the coreferences of an annotation
}

