% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp.R
\name{getOpenIE}
\alias{getOpenIE}
\title{Get OpenIE}
\usage{
getOpenIE(annotation)
}
\arguments{
\item{annotation}{an annotation object}
}
\description{
Returns a dataframe containing all OpenIE triples.
}
\examples{
getOpenIE(annoHp)

}

