% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nlp.R
\name{getParse}
\alias{getParse}
\title{Get parse tree as character vector}
\usage{
getParse(annotation)
}
\arguments{
\item{annotation}{an annotation object}
}
\description{
Returns a character vector of the parse trees.
Mostly use for visualization; the output of
\code{\link{getToken}} will generally be more
conveniant for manipulating in R.
}
\examples{
getParse(annoEtranger)
}

