% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nlp.R
\name{getToken}
\alias{getToken}
\title{Get tokens as data frame}
\usage{
getToken(annotation)
}
\arguments{
\item{annotation}{an annotation object}
}
\description{
Returns a data frame of the tokens from an annotation
object.
}
\examples{
getToken(annoEtranger)
}

