% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp.R
\name{loadXMLAnnotation}
\alias{loadXMLAnnotation}
\title{Load CoreNLP XML file}
\usage{
loadXMLAnnotation(file, encoding = "unknown")
}
\arguments{
\item{file}{connection or character string giving the file name to load}

\item{encoding}{encoding to be assumed for input strings.  It is used to mark
character strings as known to be in Latin-1 or UTF-8: it is
not used to re-encode the input. Passed to \code{readLines}.}
}
\description{
Loads a properly formated XML file output by the CoreNLP
library into an \code{annotation} object in R.
}

