% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp.R
\name{parseAnnoXML}
\alias{parseAnnoXML}
\title{Parse annotation xml}
\usage{
parseAnnoXML(xml)
}
\arguments{
\item{xml}{character vector containing the xml file from an annotation}
}
\description{
Returns an annotation object from a character vector containing
the xml. Not exported; use \code{loadXMLAnnotation} instead.
}

