% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.annotation}
\alias{plot.annotation}
\title{Print Dependencies and POS of Annotation}
\usage{
\method{plot}{annotation}(x, y = 1L, ...)
}
\arguments{
\item{x}{an annotation object}

\item{y}{sentence id}

\item{...}{other arguments passed to plot}
}
\description{
Print Dependencies and POS of Annotation
}

