% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp.R
\name{universalTagset}
\alias{universalTagset}
\title{Convert Penn TreeBank POS to Universal Tagset}
\usage{
universalTagset(pennPOS)
}
\arguments{
\item{pennPOS}{a character vector of penn tags to match}
}
\description{
Maps a character string of English Penn TreeBank
part of speech tags into the universal tagset
codes. This provides a reduced set of tags (12), and
a better cross-linguist model of speech.
}
\examples{
tok <- getToken(annoEtranger)
cbind(tok$POS,universalTagset(tok$POS))

}

