% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp.R
\name{getDependency}
\alias{getDependency}
\title{Get Dependencies}
\usage{
getDependency(annotation, type = c("CCprocessed", "basic", "collapsed"))
}
\arguments{
\item{annotation}{an annotation object}

\item{type}{the class of coreference desired}
}
\description{
Returns a data frame of the coreferences of an annotation
}
\examples{
getDependency(annoEtranger)
getDependency(annoHp)

}

