% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/nlp.R
\name{getSentiment}
\alias{getSentiment}
\title{Get Sentiment scores}
\usage{
getSentiment(annotation)
}
\arguments{
\item{annotation}{an annotation object}
}
\description{
Returns a data frame of the sentiment scores from an annotation
}

